requirements.txt
Django==4.2.11
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
